Ticket #215: Check xmf_path exists before unlink
When running shine remove from a Lustre server, the xmf file is removed
by Lustre.FileSystem.remove(). The os.unlink() will then fail with errno
2,

Closes #215.

+++ lib/Shine/Configuration/FileSystem.py
@@ -516,7 +516,9 @@ class FileSystem(object):
         if self._start_backend():
             result = self.backend.unregister_fs(self)
 
-        if not result:
+        # XMF file could have been already deleted if this is a local
+        # filesystem
+        if os.path.exists(self.xmf_path) and not result:
             os.unlink(self.xmf_path)
 
         return result
